+++ spec/lib/faraday/zipkin-tracer_spec.rb
@@ -16,32 +16,31 @@ describe ZipkinTracer::FaradayHandler do
     end
   end
 
-  # Ruby 1.8 didn't support \h
-  HEX_REGEX = RUBY_VERSION >= "1.9.2" ? '\h' : '[0-9a-fA-F]'
+  HEX_REGEX = /\A\h{16}\z/
 
-  let(:response_env) { { :status => 200 } }
-  let(:wrapped_app) { lambda{|env| ResponseObject.new(env, response_env)} }
+  let(:response_env) { { status: 200 } }
+  let(:wrapped_app) { lambda { |env| ResponseObject.new(env, response_env) } }
 
   let(:hostname) { 'service.example.com' }
   let(:host_ip) { 0x11223344 }
   let(:url_path) { '/some/path/here' }
   let(:raw_url) { "https://#{hostname}#{url_path}" }
 
-  def process(body, url, headers={})
+  def process(body, url, headers = {})
     env = {
-      :method => :post,
-      :url => url,
-      :body => body,
-      :request_headers => Faraday::Utils::Headers.new(headers),
+      method: :post,
+      url: url,
+      body: body,
+      request_headers: Faraday::Utils::Headers.new(headers),
     }
     middleware.call(env)
   end
 
-  before(:each) {
+  before do
     ::Trace.sample_rate = 0.1 # make sure initialized
-    allow(::Trace).to receive(:default_endpoint)
+    allow(::Trace).to receive(:default_endpoint).and_return(::Trace::Endpoint.new('127.0.0.1', '80', service_name))
     allow(::Trace::Endpoint).to receive(:host_to_i32).with(hostname).and_return(host_ip)
-  }
+  end
 
   shared_examples 'can make requests' do
     # helper to check host component of annotation
@@ -54,24 +53,29 @@ describe ZipkinTracer::FaradayHandler do
     def expect_tracing
       # expect SEND then RECV
       expect(::Trace).to receive(:set_rpc_name).with('post')
-      expect(::Trace).to receive(:record).with(instance_of(::Trace::BinaryAnnotation)) do |ann|
+
+      expect(middleware).to receive(:record).with(instance_of(::Trace::BinaryAnnotation)) do |ann|
         expect(ann.key).to eq('http.uri')
         expect(ann.value).to eq(url_path)
       end
-      expect(::Trace).to receive(:record).with(instance_of(::Trace::BinaryAnnotation)) do |ann|
+
+      expect(middleware).to receive(:record).with(instance_of(::Trace::BinaryAnnotation)) do |ann|
         expect(ann.key).to eq('sa')
-        expect(ann.value).to eq(1)
+        expect(ann.value).to eq('1')
         expect_host(ann.host, host_ip, service_name)
       end
-      expect(::Trace).to receive(:record).with(instance_of(::Trace::BinaryAnnotation)) do |ann|
+
+      expect(middleware).to receive(:record).with(instance_of(::Trace::BinaryAnnotation)) do |ann|
         expect(ann.key).to eq('http.status')
-        expect(ann.value).to eq(200)
+        expect(ann.value).to eq('200')
       end
-      expect(::Trace).to receive(:record).with(instance_of(::Trace::Annotation)) do |ann|
+
+      expect(middleware).to receive(:record).with(instance_of(::Trace::Annotation)) do |ann|
         expect(ann.value).to eq(::Trace::Annotation::CLIENT_SEND)
         expect_host(ann.host, '127.0.0.1', service_name)
       end.ordered
-      expect(::Trace).to receive(:record).with(instance_of(::Trace::Annotation)) do |ann|
+
+      expect(middleware).to receive(:record).with(instance_of(::Trace::Annotation)) do |ann|
         expect(ann.value).to eq(::Trace::Annotation::CLIENT_RECV)
         expect_host(ann.host, '127.0.0.1', service_name)
       end.ordered
@@ -86,10 +90,11 @@ describe ZipkinTracer::FaradayHandler do
         ::Trace.push(trace_id) do
           result = process('', url).env
         end
+
         expect(result[:request_headers]['X-B3-TraceId']).to eq('0000000000000001')
         expect(result[:request_headers]['X-B3-ParentSpanId']).to eq('0000000000000003')
         expect(result[:request_headers]['X-B3-SpanId']).not_to eq('0000000000000003')
-        expect(result[:request_headers]['X-B3-SpanId']).to match(/^#{HEX_REGEX}{16}$/)
+        expect(result[:request_headers]['X-B3-SpanId']).to match(HEX_REGEX)
         expect(result[:request_headers]['X-B3-Sampled']).to eq('true')
         expect(result[:request_headers]['X-B3-Flags']).to eq('1')
       end
@@ -107,9 +112,9 @@ describe ZipkinTracer::FaradayHandler do
       it 'generates a new ID, and sets the X-B3 request headers' do
         expect_tracing
         result = process('', url).env
-        expect(result[:request_headers]['X-B3-TraceId']).to match(/^#{HEX_REGEX}{16}$/)
-        expect(result[:request_headers]['X-B3-ParentSpanId']).to match(/^#{HEX_REGEX}{16}$/)
-        expect(result[:request_headers]['X-B3-SpanId']).to match(/^#{HEX_REGEX}{16}$/)
+        expect(result[:request_headers]['X-B3-TraceId']).to match(HEX_REGEX)
+        expect(result[:request_headers]['X-B3-ParentSpanId']).to match(HEX_REGEX)
+        expect(result[:request_headers]['X-B3-SpanId']).to match(HEX_REGEX)
         expect(result[:request_headers]['X-B3-Sampled']).to match(/(true|false)/)
         expect(result[:request_headers]['X-B3-Flags']).to match(/(1|0)/)
       end