+++ spec/lib/rack/zipkin-tracer_spec.rb
@@ -38,7 +38,7 @@ describe ZipkinTracer::RackHandler do
   end
 
   describe 'initializer' do
-    context 'configured to use kafka', :platform => :java do
+    context 'configured to use kafka', platform: :java do
       let(:zookeeper) { 'localhost:2181' }
       let(:zipkinKafkaTracer) { double('ZipkinKafkaTracer') }
 
@@ -46,7 +46,7 @@ describe ZipkinTracer::RackHandler do
         allow(::Trace::ZipkinKafkaTracer).to receive(:new) { zipkinKafkaTracer }
         expect(::Trace).to receive(:tracer=).with(zipkinKafkaTracer)
         expect(zipkinKafkaTracer).to receive(:connect)
-        middleware(app, :zookeeper => zookeeper)
+        middleware(app, zookeeper: zookeeper)
       end
     end
 
@@ -73,11 +73,33 @@ describe ZipkinTracer::RackHandler do
     describe 'sample rate initialization' do
       let(:sample_rate) { 0.42 }
       subject { middleware(app, logger: Logger.new(nil), sample_rate: sample_rate) }
+
       it 'sets the sample rate' do
         expect(::Trace).to receive(:sample_rate=).with(sample_rate)
         subject.call(mock_env)
       end
+    end
 
+    context 'no domain environment variable' do
+      before do
+        ENV['DOMAIN'] = ''
+      end
+
+      it 'sets the trace endpoint service name to the default configuration file value' do
+        expect(::Trace::Endpoint).to receive(:new).with(anything, anything, 'zipkin-tester')
+        middleware(app, service_name: 'zipkin-tester')
+      end
+    end
+
+    context 'domain environment variable initialized' do
+      before do
+        ENV['DOMAIN'] = 'zipkin-env-var-tester.example.com'
+      end
+
+      it 'sets the trace endpoint service name to the environment variable value' do
+        expect(::Trace::Endpoint).to receive(:new).with(anything, anything, 'zipkin-env-var-tester')
+        middleware(app, service_name: 'zipkin-tester')
+      end
     end
 
   end
@@ -85,6 +107,7 @@ describe ZipkinTracer::RackHandler do
   context 'Zipkin headers are passed to the middlewawre' do
     subject { middleware(app, logger: Logger.new(nil)) }
     let(:env) {mock_env(',', ZipkinTracer::RackHandler::B3_REQUIRED_HEADERS.map {|a| Hash[a, 1] }.inject(:merge))}
+
     it 'does not set the RPC method' do
       expect(::Trace).not_to receive(:set_rpc_name)
       status, headers, body = subject.call(env)
@@ -96,19 +119,19 @@ describe ZipkinTracer::RackHandler do
 
     context 'accessing a valid URL of our service' do
       before do
-        rails = double("Rails")
+        rails = double('Rails')
         allow(rails).to receive_message_chain(:application, :routes, :recognize_path).and_return({controller: 'trusmis', action: 'new'})
-        stub_const("Rails", rails)
+        stub_const('Rails', rails)
       end
       it_should_behave_like 'traces the request'
     end
 
     context 'accessing an invalid URL our our service' do
       before do
-        rails = double("Rails")
+        rails = double('Rails')
         stub_const('ActionController::RoutingError', StandardError)
         allow(rails).to receive_message_chain(:application, :routes, :recognize_path).and_raise(ActionController::RoutingError)
-        stub_const("Rails", rails)
+        stub_const('Rails', rails)
       end
 
       it 'calls the app' do
@@ -175,7 +198,7 @@ describe ZipkinTracer::RackHandler do
         ::Trace.record(::Trace::BinaryAnnotation.new('http.status', [status.to_i].pack('n'), 'I16', ::Trace.default_endpoint))
       end
     end
-    subject { middleware(app, :annotate_plugin => annotate) }
+    subject { middleware(app, annotate_plugin: annotate) }
 
     it 'traces a request with additional annotations' do
       expect(::Trace).to receive(:push).ordered
@@ -190,14 +213,14 @@ describe ZipkinTracer::RackHandler do
   end
 
   context 'configured with filter plugin that allows all' do
-    subject { middleware(app, :filter_plugin => lambda {|env| true}) }
+    subject { middleware(app, filter_plugin: lambda { |env| true }) }
 
     it_should_behave_like 'traces the request'
 
   end
 
   context 'configured with filter plugin that allows none' do
-    subject { middleware(app, :filter_plugin => lambda {|env| false}) }
+    subject { middleware(app, filter_plugin: lambda { |env| false }) }
 
     it 'does not trace the request' do
       expect(::Trace).not_to receive(:push)
@@ -210,7 +233,7 @@ describe ZipkinTracer::RackHandler do
     before(:each) { ::Trace.sample_rate = 0 }
 
     context 'configured with whitelist plugin that forces sampling' do
-      subject { middleware(app, :whitelist_plugin => lambda {|env| true}) }
+      subject { middleware(app, whitelist_plugin: lambda { |env| true }) }
 
       it 'samples the request' do
         expect(::Trace).to receive(:push) do |trace_id|
@@ -223,7 +246,7 @@ describe ZipkinTracer::RackHandler do
     end
 
     context 'configured with filter plugin that allows none' do
-      subject { middleware(app, :whitelist_plugin => lambda {|env| false}) }
+      subject { middleware(app, whitelist_plugin: lambda { |env| false }) }
 
       it 'does not sample the request' do
         allow(::Trace).to receive(:should_sample?) { false }